Ontology docs
This is not used currently, but shows how NetworkX can be used to create a visualisation of the knowledge representation graph.
NetworkX was ultimately not used, due to certain limitations in its format and a lack of effort to circumvent them (if possible). Instead, a separate list of dictionaries was created to denote the relations between the items.